pdf-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
pymupdf==1.24.10
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4

# Uncomment when ENABLE_OCR=true:
# paddleocr==2.8.1
# pdf2image==1.17.0
# Pillow==10.4.0
